server/requirements.txt
flask==3.1.0
mysql-connector-python==9.3.0
paho-mqtt==2.1.0
